==> 0_initial_graphs.ipynb <==
path = 'C:/Users/1/attribution_project/'

outcome_var = 'conv_3'
m_var = '0. Game page load'

path += 'data/'

import matplotlib.pyplot as plt
import pickle

data = pickle.load(open(path + 'data.pkl','rb'))

y = data[outcome_var]

plt.hist(y)

plt.tight_layout()
plt.savefig(path + 'y_hist.jpg')

data[m_var].hist()
plt.tight_layout()
plt.savefig(path + 'm_hist.jpg')

data.loc[data[m_var]<50, m_var].hist()
plt.tight_layout()
plt.savefig(path + 'm_hist_50.jpg')

data.loc[data[m_var]<20, m_var].hist()
plt.tight_layout()
plt.savefig(path + 'm_hist_20.jpg')
