==> 1_data_split.ipynb <==
path = 'C:/Users/1/attribution_project/'

outcome_var = 'conv_3'
m_var = '0. Game page load'

categorical_vars = ['Source', 'Medium', 'Device Category', 'placement', 'version', 'asset_type', 
                    'weekday', 'ad_id', 'ad_type', 'campaign_type']

random_state = 7

import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
import pickle
from sklearn.model_selection import train_test_split, KFold

def data_split(categorical_vars, path, m=10, random_state=7):
    
    path_in = path + 'data/'
    
    data = pickle.load(open(path_in + 'data.pkl','rb'))

    path_out = path + 'results/m' + str(m) + '/data/'

    data = data[data[m_var]>m]
    data = data.reset_index(drop=True)

    X= data.drop(outcome_var, axis=1)
    y = data[outcome_var]

    X = X.drop(m_var, axis=1)

    X['version'] = X['version'].astype('str')

    for var in categorical_vars:
        X[var] = X[var].astype('category')

    #split the data
    X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.20, random_state=random_state)
    cv = KFold(n_splits=5, random_state=random_state, shuffle=True)

    #export data

    pickle.dump(X, open(path_out + 'X.pkl','wb'))
    pickle.dump(y, open(path_out + 'y.pkl','wb'))
    pickle.dump(categorical_vars, open(path_out + 'categorical_vars.pkl','wb'))

    pickle.dump(X_train, open(path_out + 'X_train.pkl','wb'))
    pickle.dump(y_train, open(path_out + 'y_train.pkl','wb'))
    pickle.dump(X_test, open(path_out + 'X_test.pkl','wb'))
    pickle.dump(y_test, open(path_out + 'y_test.pkl','wb'))
    pickle.dump(cv, open(path_out + 'cv.pkl','wb'))
    
    plt.clf()
    plt.hist(y)
    plt.tight_layout()
    plt.savefig(path_out + 'y_hist.jpg')
    
    plt.clf()
    plt.hist(y_train)
    plt.tight_layout()
    plt.savefig(path_out + 'y_hist_train.jpg')
    
    plt.clf()
    plt.hist(y_test)
    plt.tight_layout()
    plt.savefig(path_out + 'y_hist_test.jpg')
    
    
    print(m)
    print('total:',len(y))
    print('train:',len(y_train))
    print('test:',len(y_test))

for m in [3,7,10]:
    data_split(categorical_vars, path, m=m, random_state=random_state)
